# Conflicts:
#	Example/RainSDKDemo/RainSDKDemo/Core/Services/RainSDKService.swift
#	Example/RainSDKDemo/RainSDKDemo/Presentation/PortalWithdraw/PortalWithdrawDemoViewModel.swift
#	Example/RainSDKDemo/RainSDKDemo/Presentation/TransferDemo/TransferDemoViewModel.swift
#	Sources/RainSDK/Core/RainSDKError.swift
#	Sources/RainSDK/RainSDK.swift
#	Sources/RainSDK/RainSDKManager+Helpers.swift
#	Sources/RainSDK/RainSDKManager.swift
#	Tests/RainSDKTests/SendTokenTests.swift
#	Tests/RainSDKTests/WalletAgnosticInitializationTests.swift
#	Tests/RainSDKTests/WithdrawCollateralTests.swift

+++ Example/RainSDKDemo/RainSDKDemo/Core/Services/RainSDKService.swift
@@ -161,7 +161,7 @@ class RainSDKService: ObservableObject {
     collateralProxyAddress: String,
     walletAddress: String,
     tokenAddress: String,
-    amount: Double,
+    amount: Decimal,
     decimals: Int,
     recipientAddress: String,
     nonce: BigUInt?
@@ -187,7 +187,7 @@ class RainSDKService: ObservableObject {
     contractAddress: String,
     proxyAddress: String,
     tokenAddress: String,
-    amount: Double,
+    amount: Decimal,
     decimals: Int,
     recipientAddress: String,
     expiresAt: String,
@@ -350,7 +350,7 @@ class RainSDKService: ObservableObject {
   }
 
   /// Sends native tokens (e.g. ETH, AVAX, SOL) from the current wallet.
-  func sendNativeToken(chainId: Int, to: String, amount: Double) async throws -> RainTokenTransferResult {
+  func sendNativeToken(chainId: Int, to: String, amount: Decimal) async throws -> RainTokenTransferResult {
     try await sdkManager.sendNative(chainId: chainId, to: to, amount: amount)
   }
 
@@ -360,7 +360,7 @@ class RainSDKService: ObservableObject {
     chainId: Int,
     contractAddress: String,
     to: String,
-    amount: Double,
+    amount: Decimal,
     decimals: Int? = nil
   ) async throws -> RainTokenTransferResult {
     try await sdkManager.sendToken(
@@ -381,7 +381,7 @@ class RainSDKService: ObservableObject {
     proxyAddress: String,
     tokenAddress: String,
     recipientAddress: String,
-    amount: Double,
+    amount: Decimal,
     decimals: Int,
     salt: String,
     signature: String,
