+++ Tests/RainSDKTests/Adapters/PortalAdapterTests.swift
@@ -323,7 +323,7 @@ struct PortalAdapterTests {
       expiresAt: "1735689600"
     )
 
-    let expectedFee = 20_000_000_000.0 / pow(10.0, 18.0) * 21_000.0
+    let expectedFee = Decimal(20_000_000_000.0 / pow(10.0, 18.0) * 21_000.0)
     #expect(fee == expectedFee)
   }
